split main across 6 files to help with organisation.
